[그린톡 QA 진척 업로드] UserList.jsx - 선택영역 Row로 확장
UserPromoHis.jsx
- 필터 제거
- 진행기간/참여기간 반대로 된 것 수정

UserPromoHisDetail.jsx
- target_id에 admin 상수로 박힌것 제거

UserBasicInfo.jsx
- 위 아래 수정 버튼 확인창 추가

UserPointHis.jsx
- Req Data에 Order 추가 및 기본값 Desc 지정
- 수동 입력에 오늘 날짜 디폴트 처리
- 수동 입력 날짜 인풋 박스 Disable 처리

UserQnaHis.jsx
- 배열 첫번째에 '전체'추가 && SelectValue index 에 -1
- 클릭 이벤트 영역 Row 전체로 확장

UserQnaHisDetail.jsx
- 하단 동작 하지 않는 수정 버튼 주석
- 문의 번호 UI 매칭

QnaDetail.jsx
- 수정 확인창 추가

Lnb.jsx
- 탄소 중립 랭킹 변동 내역 탭 주석 처리

ChallengeList.jsx
- 클립보드 주소 복사 이미지 주석 처리
- 클릭 이벤트 영역 Row 전체로 확장

ChallengeListDetail.jsx
- 문구 직관적으로 변경

ChallengeEdit.jsx
- 문구 직관적으로 변경
- 수정, 삭제 버튼 확인 알럿 추가
- 총 참여 회원수 모달 연결 이미지 주석

+++ src/routes/event/ChallengeEdit.jsx
@@ -93,10 +93,16 @@ export default function ChallengeEdit() {
 
   const btnEvent = {
     mod() {
-      dataSubmit();
+      if(window.confirm("수정 하시겠습니까?"))
+      {
+        dataSubmit();
+      }
     },
     del() {
-      deleteChallenge();
+      if(window.confirm("삭제 하시겠습니까?"))
+      {
+        deleteChallenge();
+      }
     },
   };
 
@@ -185,12 +191,12 @@ export default function ChallengeEdit() {
                   <div className="input_ty02 flex_left">
                     <label htmlFor="">
                       총 참여 회원 수{" "}
-                      <img
+                     {/*  <img
                         src={arrowRight}
                         alt=""
                         className="icon click"
                         onClick={() => setModalOpen(true)}
-                      />
+                      /> */}
                     </label>
                     <input
                       type="text"
@@ -246,7 +252,7 @@ export default function ChallengeEdit() {
                 </div>
                 <div className="flex_box">
                   <div className="input_ty02 flex_left mr12">
-                    <label htmlFor="">글 등록 개수(1일)</label>
+                    <label htmlFor="">달성 기준 게시글 수(1일)</label>
                     <input
                       type="text"
                       value={challengeContents.ch_stamp_board}
@@ -257,7 +263,7 @@ export default function ChallengeEdit() {
                   </div>
                   <div className="input_ty02 flex_right">
                     <label htmlFor="" className="stamp_signup">
-                      도장 적립 개수
+                    달성시 도장 적립 개수
                     </label>
                     <input
                       type="text"
@@ -270,7 +276,7 @@ export default function ChallengeEdit() {
                 </div>
                 <div className="flex_box flex_box_mb">
                   <div className="input_ty02 flex_left">
-                    <label htmlFor="">최대 적립 개수</label>
+                    <label htmlFor="">최대 도장 적립 개수</label>
                     <input
                       type="text"
                       value={challengeContents.ch_max_stamp}
@@ -317,7 +323,7 @@ export default function ChallengeEdit() {
                 </div>
                 <div className="flex_box flex_box_mb">
                   <div className="input_ty02 flex_left">
-                    <label htmlFor="">포인트 지급 조건</label>
+                    <label htmlFor="">포인트 지급 조건(도장)</label>
                     <input
                       type="text"
                       placeholder="포인트가 지급되는 도장 횟수를 숫자로 입력해주세요."
